requirements.txt
absl-py==0.7.1
appdirs==1.4.3
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
dill==0.3.0
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
googleapis-common-protos==1.6.0
graphsurgeon==0.4.1
grpcio==1.22.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
importlib-metadata==0.18
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Mako==1.0.13
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mock==3.0.5
more-itertools==7.1.0
networkx==2.3
numpy==1.16.4
onnx==1.5.0
opencv-contrib-python==4.1.0.25
packaging==19.0
Pillow==6.1.0
pluggy==0.12.0
promise==2.2.1
protobuf==3.8.0
psutil==5.6.3
py==1.8.0
pybase64==1.0.0
pycuda==2019.1.1
pyparsing==2.4.0
pytest==5.0.1
python-dateutil==2.8.0
pytools==2019.1.1
PyWavelets==1.0.3
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorflow-datasets==1.0.2
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-metadata==0.13.0
tensorrt==5.1.5.0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
tf2onnx==1.5.1
tqdm==4.32.2
typing==3.7.4
typing-extensions==3.7.4
uff==0.6.3
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.4
wrapt==1.11.2
zipp==0.5.2
